requirements.txt
adium-theme-ubuntu==0.3.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
Cython==0.27.3
dj-database-url==0.4.2
Django==1.9.13
django-heroku==0.2.0
django-rest-framework==0.1.0
django-rest-swagger==0.3.5
django-tastypie==0.14.0
djangorestframework==3.3.3
djangorestframework-jwt==1.10.0
enum34==1.1.6
ez-setup==0.9
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
jsonfield==2.0.2
mccabe==0.6.1
numpy==1.13.3
olefile==0.44
pathlib==1.0.1
pew==1.1.1
Pillow==4.3.0
pipenv==9.0.0
psycopg2==2.7.3.2
pycodestyle==2.3.1
pycurl==7.43.0
pyflakes==1.6.0
PyJWT==1.5.3
pystan==2.17.0.0
python-apt==1.1.0b1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
selenium==3.8.0
shutilwhich==1.1.0
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.1.18
tornado==4.5.2
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
xlwt==1.3.0
youtube-dl==2017.12.14
